second commit

+++ resources/views/machine/index.blade.php
@@ -5,19 +5,20 @@
 
 <div class="panel panel-default">
 	<!-- Default panel contents -->
-	<div class="panel-heading">Status das Máquinas do LAE</div>
+	<div class="panel-heading">LAE Machine Status</div>
 
 	<!-- Table -->
 	<table id="table-machine" class="table table-striped table-bordered">
 		<thead> 
 			<tr> 
 				<th>#</th> 
-				<th>Endereço IP</th> 
-				<th>Nome</th> 
-				<th>Alocado para</th>
+				<th>IP address</th> 
+				<th>Machine Name</th> 
+				<th>User Name</th>
 				<th>Status</th> 
-				<th>Ultimo Contato a (minutos) </th>
-				<th>Opções</th>
+				<th>Last Contact</th>
+				<th>Uptime</th>
+				<th>Options</th>
 			</tr> 
 		</thead>
 
@@ -35,18 +36,47 @@
 				@if ($m->user != null)
 					<td>{{$m->user->name}}</td> 
 				@else
-					<td>---</td> 
+					<td>-----</td> 
 				@endif
 
 				<td> <span id="ip{{$m->id}}" class="label label-default">loading</span> </td>
 
-				<td> {{$m->updated_at->diffInMinutes()}} </td>
+
+				@if($m->last_contact != null)
+					
+					<td> {{$m->last_contact->diffForHumans()}} </td>
+					
+				@else
+
+					<td> ----- </td>
+					
+				@endif
+
+				@if($m->uptime != null)
+					
+					<td> {{$m->uptime->diffForHumans()}} </td>
+					
+				@else
+
+					<td> ----- </td>
+					
+				@endif
+				
 				
 				<td>
-					<div class="btn-group  btn-group-xs" role="group" aria-label="...">
-						<a href="/edit/{{$m->id}}" class="btn btn-default" role="button">Editar</a>
-						<a href="/delete/{{$m->id}}" class="btn btn-default" role="button">Delete</a>
-					</div>
+					@if($m->created_by == auth()->user()->id)
+					
+						<div class="btn-group  btn-group-xs" role="group" aria-label="...">
+							<a href="/edit/{{$m->id}}" class="btn btn-default" role="button">Editar</a>
+							
+							<a href="#" data-href="/delete/{{$m->id}}" data-toggle="modal" data-target="#confirm-delete" class="btn btn-default" role="button">Delete</a>
+						</div>
+
+					@else
+
+						------
+
+					@endif
 				</td>
 
 			</tr> 
@@ -55,42 +85,52 @@
 		</tbody>
 	</table>
 </div>
+
+@include('confirmation.confirmdialog')
 {{-- <a href="/new" class="btn btn-default" role="button">Cadastrar nova maquina</a> --}}
-<a id="reload" onclick="getstatus()" class="btn btn-default" role="button">Recarregar</a>
+<a id="reload" onclick="getstatus()" class="btn btn-default" role="button">Reload</a>
 
 <p id="data"> </p>
 
 @endsection
 
 
-@if (auth()->check())
-	@section('script')
+@section('script')
+	@if (auth()->check())
 
 	<script>
-		
-		function getstatus() {
+        $('#confirm-delete').on('show.bs.modal', function(e) {
+            $(this).find('.btn-ok').attr('href', $(e.relatedTarget).data('href'));
+            
+            $('.debug-url').html('Delete URL: <strong>' + $(this).find('.btn-ok').attr('href') + '</strong>');
+        });
+    </script>
 
-			$('td').children('span').text("loading").removeClass().addClass('label label-default');
+		<script>
+			
+			function getstatus() {
 
-			$.get("{{ route('status_machine') }}", function(data, status){
-				$.each(data, function(index, value){
+				$('td').children('span').text("loading").removeClass().addClass('label label-default');
 
-					$("#ip"+index).removeClass();
+				$.get("{{ route('status_machine') }}", function(data, status){
+					$.each(data, function(index, value){
 
-					if(value == "online"){
-						$("#ip"+index).addClass("label label-success");
-						$("#ip"+index).text("online");
-					}else{
-						$("#ip"+index).addClass("label label-danger");
-						$("#ip"+index).text("offline");
-					}
+						$("#ip"+index).removeClass();
+
+						if(value == "online"){
+							$("#ip"+index).addClass("label label-success");
+							$("#ip"+index).text("online");
+						}else{
+							$("#ip"+index).addClass("label label-danger");
+							$("#ip"+index).text("offline");
+						}
+					});
 				});
-			});
-		}
+			}
 
-		getstatus();
-		// setInterval(getstatus, 30*1000);
-	</script>
-
-	@endsection
-@endif
+			getstatus();
+			// setInterval(getstatus, 30*1000);
+		</script>
+	
+	@endif
+@endsection
